+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Change Log
 
-All notable changes to the "vscode-show-migration" extension will be documented in this file.
+All notable changes to the "vscode-open-migration" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [0.1.0]
 
 - Initial release
-- Show latest migration
-- Open migration file
+- Open latest migration
+- Open migration file from list
 - Reveal migration folder in file explorer
 - Set custom migration folders for workspace